requirements.txt
matplotlib==3.8.2
numpy==1.26.2
persim==0.3.5
ripser==0.6.8
websocket-client==1.7.0